requirements.txt
eventlet==0.25.0
Flask==1.1.0
gunicorn==19.9.0
huey==2.1.0
peewee==3.9.6
pycryptodome==3.8.2
PyMySQL==0.9.3
qiniu==7.2.6
redis==3.2.1
requests==2.22.0
